+++ Localization/de-DE/Mods.PathOfTerraria.Items.hjson
@@ -1355,7 +1355,7 @@ LapisLazuliRing: {
 }
 
 ObsidianAdversaryDebug: {
-	// DisplayName: Obsidian Adversary
+	// DisplayName: Obsidian Adversary (Debug)
 	// Tooltip: Places the debug version of the Obsidian Adversary for use in Wall of Flesh mini arenas
 }
