+++ Localization/fr-FR/Mods.PathOfTerraria.SkillAugments.hjson
@@ -21,3 +21,13 @@ Critical: {
 	Name: Critique
 	Tooltip: Augmente les chances de critiques de {0}% au coût de l'augmentation de {1}% de l'utilisation de mana
 }
+
+Duplicate: {
+	// Name: Duplicate
+	// Tooltip: Allows 2 summons simultaneously at the cost of {0}% reduced damage
+}
+
+Extend: {
+	// Name: Extend
+	// Tooltip: Increases duration by {0}%
+}
