+++ Localization/pl-PL/Mods.PathOfTerraria.SkillAugments.hjson
@@ -21,3 +21,13 @@ Critical: {
 	// Name: Critical
 	// Tooltip: Increases critical chance by {0}% at the cost of {1}% increased mana usage
 }
+
+Duplicate: {
+	// Name: Duplicate
+	// Tooltip: Allows 2 summons simultaneously at the cost of {0}% reduced damage
+}
+
+Extend: {
+	// Name: Extend
+	// Tooltip: Increases duration by {0}%
+}
