+++ Localization/ru-RU/Mods.PathOfTerraria.Items.hjson
@@ -1334,7 +1334,7 @@ LapisLazuliRing: {
 }
 
 ObsidianAdversaryDebug: {
-	// DisplayName: Obsidian Adversary
+	// DisplayName: Obsidian Adversary (Debug)
 	// Tooltip: Places the debug version of the Obsidian Adversary for use in Wall of Flesh mini arenas
 }
